Acert-Backend/requirements.txt
Flask==1.1.1
gunicorn==20.0.4
google-cloud-storage==1.23.0
google-cloud-datastore==1.10.0
